+++ shared/array.h
@@ -19,4 +19,4 @@ int array_append_unique(struct array *array, const void *element);
 void array_pop(struct array *array);
 void array_free_array(struct array *array);
 void array_sort(struct array *array, int (*cmp)(const void *a, const void *b));
-int array_remove_at(struct array *array, unsigned int pos);
+int array_remove_at(struct array *array, size_t pos);
